chore: add some sane defaults to allow pids to talk
The whole network protocol is inspired by the Actor model,
each "actor" is identified by a unique PID.

By assigning a known MQTT topic structure, it is possible
to simulate a Direct P2P communication between PID. This
isn't a replacement for an RPC but one can use a two-way
direct channel to implement an RPC. Each requests contains
the address of the PID who sent the message.

BusView is hiddne behind initializers as it is meant to be
used just to debug the messages while the protocol and
interactions are being constructed.

The new PID based protocol also includes support for some
kind of broadcast messaging. Where the PID sends a message
to itself but using the `broadcast` sub-topic.